fixes #4297 feat(nimbus): move NotSet into its own component
