+++ app/experimenter/nimbus-ui/src/components/DirectoryTable/index.tsx
@@ -6,12 +6,12 @@ import React from "react";
 import { getAllExperiments_experiments } from "../../types/getAllExperiments";
 import { Link } from "@reach/router";
 import LinkExternal from "../LinkExternal";
+import NotSet from "../NotSet";
 import {
   getProposedEndDate,
   getProposedEnrollmentRange,
   humanDate,
 } from "../../lib/dateUtils";
-import { NotSet } from "../Summary";
 
 // These are all render functions for column type sin the table.
 export type ColumnComponent = React.FC<getAllExperiments_experiments>;
