+++ app/experimenter/nimbus-ui/src/components/SummaryTimeline/index.tsx
@@ -6,7 +6,7 @@ import React from "react";
 import ProgressBar from "react-bootstrap/ProgressBar";
 import pluralize from "../../lib/pluralize";
 import { getExperiment_experimentBySlug } from "../../types/getExperiment";
-import { NotSet } from "../Summary";
+import NotSet from "../NotSet";
 import { getStatus, StatusCheck } from "../../lib/experiment";
 import { humanDate } from "../../lib/dateUtils";
