+++ app/experimenter/nimbus-ui/src/components/TableAudience/index.tsx
@@ -6,7 +6,8 @@ import React from "react";
 import { getExperiment_experimentBySlug } from "../../types/getExperiment";
 import { Table } from "react-bootstrap";
 import { useConfig } from "../../hooks";
-import { displayConfigLabelOrNotSet, NotSet } from "../Summary";
+import { displayConfigLabelOrNotSet } from "../Summary";
+import NotSet from "../NotSet";
 
 type TableAudienceProps = {
   experiment: getExperiment_experimentBySlug;
